Lesson 6

+++ task_5_1.py
@@ -13,7 +13,6 @@ for i in range(9):
 print(os.listdir())
 
 for i in range(9):
-    os.remove('dir_{}'.format(i+1))
+    if os.path.exists('dir_{}'.format(i+1)):
+        os.remove('dir_{}'.format(i+1))
 print(os.listdir())
-
-
